requirements.txt
python-dotenv
newscatcherapi
openai
requests
pydub
ffmpeg
flask